gravel: move bootstrap under ctrl/nodes
